--- src/Home.js
@@ -1,11 +1,0 @@
-import React from 'react';
-import './Home.css';
-import StickySidebar from "./components/StickySidebar/StickySidebar";
-
-function Home() {
-  return (
-   <StickySidebar/>
-  );
-}
-
-export default Home;
