+++ src/home/styles.js
@@ -1,0 +1,17 @@
+import styled from "styled-components";
+import {Menu} from "semantic-ui-react";
+
+export const StyledMenu = styled(Menu)`
+  &&& {
+    width: 280px !important;
+    background-color: #DB1793 !important;
+    border-right: none;
+    text-align: center;
+  }
+`;
+
+export const StyledArticle = styled.div`
+  &&& {
+    margin-left: 300px;
+  }
+`;
