+++ src/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
-import Home from './Home';
+import Home from './home/Home';
 import 'semantic-ui-css/semantic.min.css'
 
 ReactDOM.render(<Home/>, document.getElementById('root'));
